chapter7 done and bugfix
