+++ netflix-clone/src/components/Banner.scss
