+++ netflix-clone/src/components/Nav.scss
